+++ src/main/java/com/pepper/project/cm/community/service/CommunityServiceImpl.java
@@ -43,7 +43,7 @@ public class CommunityServiceImpl implements ICommunityService {
 
         List<Community> list = communityDao.selectCommunityList(community);
         for (Community hos:list) {
-            Area area = areaDao.selectAreaById(Integer.valueOf(hos.getAreaId()));
+            Area area = areaDao.selectAreaById(Long.valueOf(hos.getAreaId()));
             hos.setAreaName(area.getAreaName());
         }
         return list;
