+++ src/main/java/com/pepper/project/csc/area/mapper/AreaMapper.java
@@ -5,7 +5,7 @@ import com.pepper.project.csc.area.domain.Area;
 import java.util.List;
 
 public interface AreaMapper {
-    Area selectAreaById(Integer id);
+    Area selectAreaById(Long id);
     List<Area> selectAreaList(Area area);
 
     /**
