+++ src/main/java/com/pepper/project/csc/area/service/AreaServiceImpl.java
@@ -44,7 +44,7 @@ public class AreaServiceImpl implements IAreaService {
     }
 
     @Override
-    public Area selectAreaById(Integer id) {
+    public Area selectAreaById(Long id) {
         return areaDao.selectAreaById(id);
     }
