+++ src/main/java/com/pepper/project/pm/property/service/PropertyServiceImpl.java
@@ -37,7 +37,7 @@ public class PropertyServiceImpl implements IPropertyService {
 
         List<Property> list = propertyDao.selectPropertyList(property);
         for (Property hos:list) {
-            Area area = areaDao.selectAreaById(Integer.valueOf(hos.getAreaId()));
+            Area area = areaDao.selectAreaById(Long.valueOf(hos.getAreaId()));
             hos.setAreaName(area.getAreaName());
         }
         return list;
